backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.1.4
chromadb>=0.5.0
python-dotenv==1.0.1
requests==2.31.0
playwright
